requirements.txt
boto3>=1.26.0
pandas>=2.0.0
pyarrow>=14.0.1
pyspark>=3.4.0
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-emr-containers>=2.0.0
aws-cdk.aws-glue>=2.0.0
aws-cdk.aws-athena>=2.0.0
aws-cdk.aws-quicksight>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
numpy>=1.24.0 